+++ Cassandra.ThriftClient/Abstractions/TimeBasedColumn.cs
@@ -1,0 +1,12 @@
+using SkbKontur.Cassandra.TimeBasedUuid;
+
+namespace SKBKontur.Cassandra.CassandraClient.Abstractions
+{
+    public class TimeBasedColumn
+    {
+        public TimeGuid Name { get; set; }
+        public byte[] Value { get; set; }
+        public long? Timestamp { get; set; }
+        public int? Ttl { get; set; }
+    }
+}